move allToolsList with scroll into separate file
